requirements.txt
psutil==5.7.0
pyserial==3.4
